use webview to implement sendD2CMessageRepeatedly(constant message)
