--- src/marketplace/assets/index.html
@@ -1,105 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-    <link rel="stylesheet" href="{{root}}/src/marketplace/assets/main.css">
-    <link rel="stylesheet" href="./main.css">
-</head>
-
-<body>
-    <div id="app" data-endpoint="{{endpoint}}" v-cloak>
-        <div v-for="item in list">
-            <p> {{ item }} </p>
-        </div>
-        <p> name@@@{{ name }}@@@ </p>
-    </div>
-    <!-- <div id="app" data-endpoint="{{endpoint}}" v-cloak>
-        <div id="content">
-            <div id="main-content">
-                <div v-if="!errorMessageInitialization">
-                    <header>
-                        <div id="title-area">
-                            <div v-if="!searchInput">
-                                <h1>Azure IoT Edge Module Marketplace ({{ this.filteredModules.length }})</h1>
-                            </div>
-                            <div v-else>
-                                <h1>Result for {{ searchInput }} ({{ this.filteredModules.length }})</h1>
-                            </div>
-                            <input id="search-box" v-model="searchInput" placeholder="Search">
-                        </div>
-                    </header>
-                    <div class="modules">
-                        <div v-for="module in filteredModules" class="module" :class="{selected:selectedModule.id == module.id}"
-                            v-on:click="showModule(module)">
-                            <div class="image">
-                                <img :src="module.icon">
-                            </div>
-                            <div class="info-area">
-                                <div class="title">{{ module.displayName }}</div>
-                                <div class="publisher">By {{ module.publisherDisplayName }}</div>
-                                <div class="description">{{ module.summary }}</div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-                <div v-else id="error-initialization">
-                    Fail to fetch IoT Edge modules from Azure Marketplace. Please retry. <br />
-                    Error message: {{ errorMessageInitialization }}
-                </div>
-            </div>
-            <div id="module-detail" v-if="selectedModule.id">
-                <div id="module-detail-container">
-                    <div class="info-item">
-                        <div class="title"> {{ selectedModule.displayName }} </div>
-                        <div class="publisher">By {{ selectedModule.publisherDisplayName }}</div>
-                        <div class="description" v-html="selectedModule.description"></div>
-                        <div class="inputs">
-                            <div class="input">
-                                <label>Module Name: </label>
-                                <input v-model="moduleName">
-                            </div>
-                            <div class="error" v-if="errorMessageModuleName">
-                                {{ errorMessageModuleName }}
-                            </div>
-                            <div class="input">
-                                <label>Tags: </label>
-                                <select v-model="selectedTag">
-                                    <option v-for="tag in selectedModule.metadata.tagsOrDigests">{{ tag }}</option>
-                                </select>
-                            </div>
-                        </div>
-                        <div class="action">
-                            <button v-on:click="importModule()">Import</button>
-                            <button v-on:click="selectedModule.id=''">Cancel</button>
-                        </div>
-                        <div class="term-of-use">
-                            I agree to the provider's <a :href="selectedModule.legalTermsUri">Term of Use</a> and <a :href="selectedModule.privacyPolicyUri">Privacy
-                                Policy</a> and understand that the rights to use this product do not come from Microsoft,
-                            unless Microsoft is the provider. Use of Azure Marketplace is governed by separate <a href="https://aka.ms/azuremarketplaceterms">terms</a>.
-                        </div>
-                    </div>
-                    <div class="learn-more-item">
-                        <div v-if="selectedModule.links && selectedModule.links.length > 0">
-                            <div class="learn-more"> Learn more </div>
-                            <div class="learn-more-links">
-                                <div v-for="link in selectedModule.links" class="link">
-                                    <a :href="link.uri">{{ link.displayName }}</a>
-                                </div>
-                            </div>
-                        </div>
-                    </div>
-                </div>
-            </div>
-        </div>
-        <div id="sidebar">
-        </div>
-    </div> -->
-    <script type="text/javascript" src="{{root}}/src/marketplace/assets/vue.min.js"></script>
-    <script type="text/javascript" src="./vue.min.js"></script>
-    <script type="text/javascript" src="{{root}}/src/marketplace/assets/axios.min.js"></script>
-    <script type="text/javascript" src="./axios.min.js"></script>
-    <script type="text/javascript" src="{{root}}/src/marketplace/assets/main.js"></script>
-    <script type="text/javascript" src="./main.js"></script>
-</body>
-
-</html>
