+++ src/simulatorwebview/assets/index.html
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+    <link rel="stylesheet" href="{{root}}/src/simulatorwebview/assets/main.css">
+    <link rel="stylesheet" href="./main.css">
+    <link rel="stylesheet" href="{{root}}/src/simulatorwebview/assets/iview.css">
+    <link rel="stylesheet" href="./iview.css">
+</head>
+
+<body>
+    <div id="app" data-endpoint="{{endpoint}}" v-cloak>
+        <i-select
+        v-model="inputDevice"
+        multiple
+        filterable
+        placeholder="Please select the device(s) you want to send message to..."
+        >
+            <i-option
+            v-for="item in inputDeviceList"
+            :value="item.connectionString"
+            :key="item.connectionString"
+            >
+                {{ item.label }}
+            </i-option>
+        </i-select>
+        <i-input v-model="msg" placeholder="Enter msg..."></i-input>
+        <i-input v-model="times" placeholder="Enter times..."></i-input>
+        <i-input v-model="interval" placeholder="Enter interval..."></i-input>
+        <i-input v-model="result"></i-input>
+        <i-button @click="send()" type="success" long> send </i-button>
+    </div>
+    <script type="text/javascript" src="{{root}}/src/simulatorwebview/assets/vue.min.js"></script>
+    <script type="text/javascript" src="./vue.min.js"></script>
+    <script type="text/javascript" src="{{root}}/src/simulatorwebview/assets/axios.min.js"></script>
+    <script type="text/javascript" src="./axios.min.js"></script>
+    <script type="text/javascript" src="{{root}}/src/simulatorwebview/assets/main.js"></script>
+    <script type="text/javascript" src="./main.js"></script>
+    <script type="text/javascript" src="{{root}}/src/simulatorwebview/assets/iview.min.js"></script>
+    <script type="text/javascript" src="./iview.min.js"></script>
+</body>
+
+</html>
